ride-sharing-backend/services/trip-service/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic-settings
kafka-python
redis
httpx
websockets
python-socketio
email-validator
